app-redimensionador/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
jinja2==3.1.4
python-multipart==0.0.12
pillow>=10.0.0
